Add a method to calculate the Fibonacci number.

+++ src/Fibonacci.java
@@ -2,8 +2,8 @@ public class Fibonacci {
     public static void main(String[] args) {
         System.out.println(fibonacciEffective(100)); // add your number
     }
-    private static int fibonacciEffective (int n) {
-        int [] arr = new int[n +1];
+    private static long fibonacciEffective (int n) {
+        long [] arr = new long[n +1];
 
         arr[0] = 0;
         arr[1] = 1;
